Move ScopedEdit into own file
